requirements.txt
requests==2.18.4
python_dateutil==2.8.1
pipreqs
